# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ALSS-LEPC Adaptive Learning API",
    "description": "Adaptive Learning System with SAINT KT + D3QN RL",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/session/{session_id}/graph": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Knowledge Graph",
        "operationId": "get_knowledge_graph_api_session__session_id__graph_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeGraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/mastery-matrix": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Mastery Matrix",
        "operationId": "get_mastery_matrix_api_session__session_id__mastery_matrix_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasteryMatrixResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/concept/{concept_id}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Concept Detail",
        "operationId": "get_concept_detail_api_session__session_id__concept__concept_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "concept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Concept Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConceptDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConceptDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          },
          "mastery": {
            "type": "number",
            "title": "Mastery"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "bloom_mastery": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Bloom Mastery"
          },
          "prerequisites": {
            "items": {
              "$ref": "#/components/schemas/ConceptPrereq"
            },
            "type": "array",
            "title": "Prerequisites"
          },
          "dependents": {
            "items": {
              "$ref": "#/components/schemas/ConceptPrereq"
            },
            "type": "array",
            "title": "Dependents"
          },
          "visited": {
            "type": "boolean",
            "title": "Visited"
          },
          "visit_count": {
            "type": "integer",
            "title": "Visit Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "definition",
          "mastery",
          "status",
          "bloom_mastery",
          "prerequisites",
          "dependents",
          "visited",
          "visit_count"
        ],
        "title": "ConceptDetailResponse"
      },
      "ConceptPrereq": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mastery": {
            "type": "number",
            "title": "Mastery"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "mastery"
        ],
        "title": "ConceptPrereq"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KnowledgeEdgeResponse": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target"
        ],
        "title": "KnowledgeEdgeResponse"
      },
      "KnowledgeGraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeNodeResponse"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "title": "KnowledgeGraphResponse"
      },
      "KnowledgeNodeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mastery": {
            "type": "number",
            "title": "Mastery"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "visited": {
            "type": "boolean",
            "title": "Visited"
          }
        },
        "type": "object",
        "required": [
          "id",
          "index",
          "name",
          "mastery",
          "status",
          "visited"
        ],
        "title": "KnowledgeNodeResponse"
      },
      "MasteryMatrixResponse": {
        "properties": {
          "matrix": {
            "items": {
              "$ref": "#/components/schemas/MasteryRow"
            },
            "type": "array",
            "title": "Matrix"
          },
          "bloom_labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bloom Labels"
          }
        },
        "type": "object",
        "required": [
          "matrix",
          "bloom_labels"
        ],
        "title": "MasteryMatrixResponse"
      },
      "MasteryRow": {
        "properties": {
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "concept_name": {
            "type": "string",
            "title": "Concept Name"
          },
          "bloom_levels": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Bloom Levels"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "concept_name",
          "bloom_levels"
        ],
        "title": "MasteryRow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "XUserIdHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Authenticated user id forwarded by the frontend proxy."
      },
      "XServiceTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-service-token",
        "description": "Shared internal token used by the frontend proxy when calling the backend API."
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment behind the current API host."
    }
  ],
  "security": [
    {
      "XUserIdHeader": [],
      "XServiceTokenHeader": []
    }
  ]
}